Error Fix for All Lifestage Category & Sciencebase
This fixes the All Lifestages error: the sum of the other lifestages was
not equal to the All Lifestages category.
The Sciencebase link to the data/metadata has been changed to reflect a
single landing page for all subsequent data releases - this should
result in not having to change the link every year.  The Shiny App was
updated on 4-4-2017.